requirements.txt
streamlit
pandas
requests
plotly
python-dotenv
cohere
